requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
chromadb
pypdf
python-dotenv